Reworked stats and hits
We do collect IP now. Since we have no way of contacting users, this might offer a clue to come back to them if we have to implement hard changes that needs to be communicated. If this ever happens, we'll get people on board with accounts and get rid of this.